Add Pierre::Store::Importer
This imports CSV files

Needed the functionality, so I built it -
but I'm not completely confident in the implementation.

Would love to chat about it with someone! :D

---

Currently, only importing allowed is CSV, and context is auto-assigned
to :en

I think, ideally - we could have multiple types of Importers
(json/csv/yaml/etc) - each of which inherit from the base Importer.

Then you could pass in an option to specificy which type (or maybe we
just auto figure it out based on extension?)